Fix bug in android speech_commands average score calculation, add unit test
Previous logic added a reference to the result array to the Deque (as opposed
to the value of the result array). When a new result was added to the deque,
all previous results were updated to this value (because they were references)
meaning the averaging operation was performed on a deque of identical arrays -
effectively just returning the final value added to the Deque, rather than an
average of previous results.

The code change extracts this updating logic to a function which is independently
testable, and modifies the logic to add the array value rather than a reference.