Mode Maker: DataLoader refactor the `gen_dataset` internal API.
Instead of exposing `dataset` attribute directly, this CL wraps it in a method.
This is a mechanical move. No functionality should be impacted. In most of the cases (outside the context of task), `dataload.dataset` could be replaced by `gen_dataset()`

This is the first step for making the Dataset doing Lazy load.

PiperOrigin-RevId: 340396087