Model Maker: Refactor audio_dataloader class.
At the momement, we first apply map then zip on the dataset, this will cause data consistency issues once we apply filters on the dataset and potentially use unordered_map to accelerate the processing.

This CL switch to zip then map to keep label and example in sync regardless of how the data is processed. It also aligns with how image preprocessing is being done.

This refactor should have no impact on the functionality.

PiperOrigin-RevId: 345171679