Model Maker Audio refactor 1: Stop using exact DataLoader length for training.
We're unable to get the exact DataLoader size for Browser FFT and Yamnet preprocessing. We should stop relying on the dataset size. len(data_loader) is used for both `training` and `split`. This CL removes the need of exact size for training only. There will be a follow up CL for `split`.

Detailed changes:

- gen_dataset method has been copied to Audio DataLoader, except that the train_ds is no longer repeated.
- Training the model no longer requires steps_per_epochs and validation_steps. No data will be lost since, by default we keep remainder when batching the data.

This only works for audio model for now, but we should be able to extend the idea to other tasks if needed.

PiperOrigin-RevId: 354302981