Model Maker Audio refactor 3: No longer need to create cache.
Creating cache is expensive and increases (doubles) the disk usage. Previously we had to create cache in order to accurately count how many samples are there in the dataset. Since we no longer need this information when training the model, we can simplify the dataset creation process by putting resample and cut into tf.dataset

PiperOrigin-RevId: 354497190