added class_weight for imbalanced datasets
- added class_weight to create and train method in image_classifier.py
- added class_weight to train_model method in train_image_classifier.py
All this was done so that the class_weight can be added to the model.fit() because Keras Models already have this as an argument.

I was not able to actually test if this works because I can't build from source when I tried from the readme instructions for model maker (I'm not sure why I can't build). So, these are the changes I think will work. Keras Model already has class_weight as an argument in the .fit() so I just tried to change the code so that a class_weight can be added from the .create() of the model_maker for image classification.