Disable saved model lowering during keras model conversion
TFLite team is going to lower the keras model to the corresponding saved model
for the new feature additions, introduced in saved model conversion. However,
the new path will generate a slightly different model in terms of tensor
ordering and tensor name due to the saved model serialization process. To
avoid the breakages, we suggest disabling saved model lowering explicitly.

PiperOrigin-RevId: 384656358