Better support for alpha channels
Tweaking the unet_generator and discriminator functions to allow for output channels other than 3. unet_generator had a bug in that the Input shape was not respecting the output_channel parameter and discriminator did not support this functionality.

With these changes the CycleGAN tutorial can be modified by learners to support .png files more easily.