Xcode 13 settings and pod install for M1 note
I've made minor updates to the Signing and Bundle Identifier setting instructions to reflect the current state of settings locations in Xcode 13. (Not sure if the tabs were different in earlier versions.)
I've also added a second note after the `pod install` step. Not sure if linking to some Stack Overflow post like this is allowed, but I had a lot of trouble getting `pod install` to run successfully on an M1 Mac. Wanted to add something here to steer others in a hopefully helpful direction.