Fix Pascal Conversion
Data exported from CVAT to the Pascal VOC format does not have the truncated or pose fields. It also specifies coordinates using float values instead of ints.

An earlier part of the code casts `obj['bndbox']['xmin']` to a float, so it seems valid to do this in the second place that accesses `obj['bndbox']['xmin']`, and just round this value. The behavior should be identical when values are ints. 

The behavior is identical when difficult and truncated fields are specified. However, this change also makes this code work correctly for files that do not contain these fields.